+++ src/Http/Requests/FormRequest.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * Copyright 2021 Cloud Creativity Limited
+ * Copyright 2022 Cloud Creativity Limited
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,13 +21,11 @@ namespace LaravelJsonApi\Laravel\Http\Requests;
 
 use Illuminate\Auth\AuthenticationException;
 use Illuminate\Contracts\Auth\Guard;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Foundation\Http\FormRequest as BaseFormRequest;
 use Illuminate\Support\Str;
 use LaravelJsonApi\Contracts\Schema\Schema;
 use LaravelJsonApi\Core\JsonApiService;
 use LaravelJsonApi\Validation\Factory as ValidationFactory;
-use LogicException;
 
 class FormRequest extends BaseFormRequest
 {
@@ -229,36 +227,6 @@ class FormRequest extends BaseFormRequest
     public function schema(): Schema
     {
         return $this->jsonApi()->route()->schema();
-    }
-
-    /**
-     * Get the model that the request relates to, if the URL has a resource id.
-     *
-     * @return Model|object|null
-     */
-    public function model(): ?object
-    {
-        $route = $this->jsonApi()->route();
-
-        if ($route->hasResourceId()) {
-            return $route->model();
-        }
-
-        return null;
-    }
-
-    /**
-     * Get the model that the request relates to, or fail if there is none.
-     *
-     * @return Model|object
-     */
-    public function modelOrFail(): object
-    {
-        if ($model = $this->model()) {
-            return $model;
-        }
-
-        throw new LogicException('No model exists for this route.');
     }
 
     /**